pyproject.toml
[project]
name = "ensemble"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.115.8",
    "python-decouple>=3.8",
    "sqlmodel>=0.0.22",
    "uvicorn>=0.34.0",
    "websockets>=15.0",
]
